to make simple

+++ inc/in.h
@@ -7,4 +7,4 @@ struct sockaddr_in {
     struct in_addr   sin_addr;     // see struct in_addr, below
     char             sin_zero[8];  // zero this if you want to
 };
-#define htons(a) (((a)>>8) | ((a)<<8))
+#define htons(a) (unsigned short)(((a)>>8) | ((a)<<8))
